+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.6.3] 2022-05-18
+## [0.6.3] 2022-05-2020
 
 ### Added
 
 - Add support of prefab's links within scene's assets.
+
+- Add a new component **CPrefabLinkInfoComponent** to store in-editor information about instantiated prefabs. 
 
 ### Fixed